requirements.txt
Django==2.1.5
django-appconf==1.0.2
django-debug-toolbar==1.11
django-filter==2.0.0
django-imagekit==4.0.2
django-reset-migrations==0.3.1
djangorestframework==3.9.0
pilkit==2.0
Pillow==5.4.1
pytz==2018.9
six==1.12.0
sqlparse==0.2.4
